email-user-to-update-profile
We added in a rake task that can be called via cron which will
iterate through all of the users and check the fields in their
profile to see if they need updating, and then send that user an email
that lists the fields that they need to update.

We set this to run via cron once a month on the first of the month.

Authors:
Craig Hokanson
Tim Zaitsev
James Ricks