Initial implementation of Teaching Assistant role.
(In the current system, when we want to add a TA, we list them as a course owner. We'd like to separate this out into its own role.)

As a course instructor, I want to add a TA to a course, so (s)he is listed as a TA on the main course page (under Course Tools) and can receive/send emails about the course.
A TA can be any registered students (provided the student is not currently enrolled in that course). I may assign more than one TA to a course. Only course instructors or admins can assign a TA to a course. By default, anything that a course faculty can do, a TA can do too. (Anywhere the code does faculty.include?, it should also work for TAs.)